Add 2 function and modify the screen calculation
Implement the ft_open_link function that is open
the link passed to it as argument in the default
browser by using system() that execute a command
from c code into terminal and terminal handle
passing link to browser,
Also the ft_opening_message that create its window
and print that link is opened in browser to notice
the user.
Extra thing is that i modify the window calculation
from dividing the hole standard screen by 2, instead
of that i'd use the ratio by multiply the size by
0.50 to use the half screen, that is more readable
and scalable.