Implement function to handle window resizing
Since the the terminal window can be resized during
the tool running so the function is passed to
the signal standard C function that use the handle
function to work with that resizing.
(The signal() from C standard is receiving the signal
from the OS and call our handle function to do
the call up in that case) so our function is clear
the screen and quite the mode, and redraw the
standard screen with the menu.